Columns from integer to bigInteger

+++ db/migrations/dev/20210411003659_build-job-table.js
@@ -2,9 +2,9 @@ exports.up = function (knex) {
   return knex.schema.createTable("jobs", (table) => {
     table.increments("jobId");
     table.text("name", 128);
-    table.integer("startTime");
-    table.integer("endTime");
-    table.integer("length");
+    table.bigInteger("startTime");
+    table.bigInteger("endTime");
+    table.bigInteger("length");
     table.boolean("isFinished").defaultTo(false);
   });
 };
